requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
appier==1.34.12
attrs==25.4.0
binance-api==0.1.2
blinker==1.9.0
build==1.3.0
cachelib==0.13.0
ccxt==4.5.11
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
dash==3.2.0
dash-bootstrap-components==2.0.4
dash-extensions==2.0.4
dataclass-wizard==0.35.1
dateparser==1.2.2
EditorConfig==0.17.1
fastapi==0.120.0
Flask==3.1.2
Flask-Caching==2.3.1
flask-cors==6.0.1
fonttools==4.60.1
frozenlist==1.8.0
h11==0.16.0
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsbeautifier==1.15.4
kiwisolver==1.4.9
llvmlite==0.44.0
MarkupSafe==3.0.3
matplotlib==3.10.7
more-itertools==10.8.0
multidict==6.7.0
narwhals==2.9.0
nest-asyncio==1.6.0
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.3.3
pandas-ta==0.4.71b0
pillow==12.0.0
plotly==6.3.1
propcache==0.4.1
psycopg2==2.9.11
psycopg2-binary==2.9.11
pycares==4.11.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.3
pydantic_core==2.41.4
pyparsing==3.2.5
pyproject_hooks==1.2.0
python-binance==1.0.30
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
regex==2025.9.18
requests==2.32.5
retrying==1.4.2
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
ta==0.11.0
TA-Lib==0.6.8
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.22.0
zipp==3.23.0
